requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
PyJWT==2.1.0
python-dotenv==0.18.0
semver==2.13.0
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
flask_jwt_extended
